+++ pkg/scalers/influxdb_scaler.go
@@ -122,6 +122,7 @@ func parseInfluxDBMetadata(config *ScalerConfig) (*influxDBMetadata, error) {
 		return nil, fmt.Errorf("no server url given")
 	}
 
+	// FIXME: DEPRECATED to be removed in v2.12
 	if val, ok := config.TriggerMetadata["metricName"]; ok {
 		metricName = kedautil.NormalizeString(fmt.Sprintf("influxdb-%s", val))
 	} else {